backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-q2==1.6.2
psycopg[binary]>=3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
